refactor: move criterion functions
