+++ pyMixtComp/python/pyMixtComp/plot/barplot.py
@@ -1,72 +1,6 @@
-import numpy as np
 import seaborn as sns
 
-
-def compute_discriminative_power_variable(res, class_id=None):
-    """ Compute the discriminative power of variables
-
-    Parameters
-    ----------
-    res : dict
-        output of multi_run_pmc_pool
-    class_id : int, optional
-        Class number (0, ...,  n_components - 1) of class name, by default None.
-        If None, it returns the discriminative power of variables globally otherwise it returns the discriminative power
-        of variables in the given class
-
-    Returns
-    -------
-    Series
-        Discriminative power of variables
-
-    Notes
-    -----
-    The discriminative power of variable j is defined by:
-    .. math::
-    1 - \sum_{k=1}^K sum_{i=1}^n P(Z_i=k|x_{ij}) \log(P(Z_i=k|x_{ij})) / (n*\log(K))
-
-    The discriminative power of variable j in class k is defined by:
-    .. math::
-    1 - (sum_{i=1}^n P(Z_i!=k|x_{ij}) \log(P(Z_i!=k|x_{ij})) + sum_{i=1}^n P(Z_i=k|x_{ij}) \log(P(Z_i=k|x_{ij}))) / (n*\log(2))
-
-    A high value (close to one) means that the variable is highly discriminating.
-    A low value (close to zero) means that the variable is poorly discriminating.
-    """
-    if class_id is None:
-        return 1 - res["mixture"]["IDClass"].sum(axis=0)
-    else:
-        K = len(res["mixture"]["IDClass"])
-
-        entropy_KJ = res["mixture"]["IDClass"].loc["k: " + str(class_id)] * np.log(K)
-        entropy_KJ_bar = res["mixture"]["IDClassBar"].loc["k: " + str(class_id)] * np.log(K)
-
-        return 1 - (entropy_KJ + entropy_KJ_bar) / np.log(2)
-
-
-def compute_discriminative_power_class(res):
-    """ Compute the discriminative power of classes
-
-    Parameters
-    ----------
-    res : dict
-        output of multi_run_pmc_pool
-
-    Returns
-    -------
-    Series
-        discriminative power of classes
-
-    Notes
-    -----
-    The discriminative power of classes is defined by:
-    .. math::
-    1 + \sum_{i=1}^n P(Z_i=k|x_i) \log(P(Z_i=k|x_i)) / (n*\exp(-1))}
-
-    A high value (close to one) means that the class is highly discriminating.
-    A low value (close to zero) means that the class is poorly discriminating.
-    """
-    tik = res["variable"]["data"]["z_class"]["stat"]
-    return 1 + np.log(tik ** tik).mean(axis=0)/np.exp(-1)
+from pyMixtComp.utils.criterion import compute_discriminative_power_class, compute_discriminative_power_variable
 
 
 def plot_discriminative_power_variable(res, class_id=None):
