+++ pyMixtComp/python/pyMixtComp/utils/__init__.py
@@ -1,3 +1,5 @@
 # flake8: noqa
 from .getter import get_tik, get_partition, get_param
 from .convert import create_functional, convert_functional, refactor_categorical
+from .criterion import compute_variable_similarity, compute_class_similarity, \
+    compute_discriminative_power_variable, compute_discriminative_power_class
